Rename evaluateVelocity function
